Moved code to different file for pushing into kinesis
