+++ lib/stream.js
@@ -1,19 +1,22 @@
+const avro = require('avsc');
 const AWS = require('aws-sdk')
-const _ = require('highland')
 const crypto = require('crypto')
-
 const kinesis = new AWS.Kinesis({region: 'us-east-1'})
 
-exports.postBibsStream = function(bibs,callback){
-  var params = {
-    Data: JSON.stringify(bibs), /* required */
-    PartitionKey: crypto.randomBytes(20).toString('hex').toString(), /* required */
-    StreamName: 'testBibRetrieval', /* required */
-  }
-  kinesis.putRecord(params, function (err, data) {
-    if (err) console.log(err, err.stack) // an error occurred
-    else     console.log(data)           // successful response
-    cb(null,data)
+//send data to kinesis Stream
+exports.postResourcesToStream = function(resource, stream){
+  return new Promise((resolve, reject) => {
+    const type = avro.infer(resource);
+    const resource_in_avro_format = type.toBuffer(resource);
+    var params = {
+      Data: resource_in_avro_format, // required 
+      PartitionKey: crypto.randomBytes(20).toString('hex').toString(), // required
+      StreamName: stream, // required 
+    }
+    kinesis.putRecord(params, function (err, data) {
+      console.log('wrote to kinesis: ', data)
+      if (err) reject(err)
+      else resolve(data)
+    })
   })
-
 }
